ci: update rfc2307bis.ldif
The version I found seemed to be based on an earlier version of the
draft. Crucially, it didn't include the ORDERING property on the
shadowExpire attribute (added in draft-02). This prevented the
expiredmembers.php script from finding expired memberships.

The replacement schema is an LDIF conversion of the following:

https://github.com/jtyr/rfc2307bis/blob/master/rfc2307bis.schema

The version being replaced also included a copy of the namedobject
schema, which this one doesn't. So we need to add that one manually.