lib, www: switch to using the single displayName field everywhere rather than givenName/sn
The LDAP schema still requires that person objects have an 'sn'
attribute, so we set it to '_' (which the previous code did if the user
had not set a surname).

No data migration is necessary, since we were already setting the
'displayName` attribute. If the full name on any person object is
changed, the 'givenName' attribute will be removed and 'sn' set to '_'.